Extract package managers from package-cache
In order to bring npm and yarn (and bower as a bonus) invocation into
a single place, I’m starting by extrating them from the place where
they are mostly used, `package-cache`. This will allow us to use this
module as the central point for functionalities related with abstracting
which package manager is running, focusing on the goal.
