+++ lib/tasks/install-blueprint.js
@@ -171,6 +171,7 @@ class InstallBlueprintTask extends Task {
 
     this._copyNpmrc(cwd);
 
+    // TODO: Teach this class about yarn
     return execa('npm', ['install', '--no-package-lock', '--save', name], { cwd });
   }
