+++ lib/utilities/command-generator.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const execa = require('../../lib/utilities/execa');
+const execa = require('./execa');
 
 /**
  * A simple tool to make behavior and API consistent between commands
